+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardGeneralStep/index.js
@@ -1,28 +1,6 @@
-import React from 'react';
 import { connect } from 'react-redux';
-import { Field, reduxForm } from 'redux-form';
-import { required } from 'redux-form-validators';
-import { Form } from 'patternfly-react';
-import { FormField } from '../../../formUtilities/FormField';
-
-const MappingWizardGeneralStep = props => (
-  <Form className="form-horizontal">
-    <Field
-      name="name"
-      label={__('Name')}
-      required
-      component={FormField}
-      validate={[required({ msg: __('Required') })]}
-      type="text"
-    />
-    <Field
-      name="description"
-      label={__('Description')}
-      component={FormField}
-      type="textarea"
-    />
-  </Form>
-);
+import { reduxForm } from 'redux-form';
+import MappingWizardGeneralStep from './MappingWizardGeneralStep';
 
 // Decorate with reduxForm(). It will read the initialValues prop provided by connect()
 const MappingWizardGeneralStepForm = reduxForm({
